+++ src/main/java/umc/product/domain/checklist/service/ChecklistCommandService.java
@@ -1,7 +1,7 @@
 package umc.product.domain.checklist.service;
 
-import umc.product.domain.study.dto.common.request.StudyChecklistListRequest;
-import umc.product.domain.study.dto.common.response.StudyCommonResponse;
+import umc.product.domain.study.dto.request.member.StudyChecklistListRequest;
+import umc.product.domain.study.dto.response.member.StudyCommonResponse;
 import umc.product.domain.study.entity.StudyMember;
 
 import java.util.List;
