--- src/main/java/umc/product/domain/study/repository/StudyRepository.java
@@ -1,7 +1,0 @@
-package umc.product.domain.study.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import umc.product.domain.study.entity.Study;
-
-public interface StudyRepository extends JpaRepository<Study, Long>, StudyCustomRepository {
-}
