+++ src/main/java/umc/product/domain/study/repository/member/StudyRepository.java
@@ -1,0 +1,7 @@
+package umc.product.domain.study.repository.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import umc.product.domain.study.entity.Study;
+
+public interface StudyRepository extends JpaRepository<Study, Long>, StudyCustomRepository {
+}
