--- src/main/java/umc/product/domain/study/service/StudyMemberQueryService.java
@@ -1,11 +1,0 @@
-package umc.product.domain.study.service;
-
-import umc.product.domain.member.entity.Member;
-import umc.product.domain.study.entity.StudyMember;
-
-public interface StudyMemberQueryService {
-
-    StudyMember getStudyMember(Member member, Long studyId);
-
-    StudyMember getStudyMemberFetch(Member member, Long studyId);
-}
