+++ src/main/java/umc/product/domain/study/service/member/StudyAttendanceQueryService.java
@@ -1,0 +1,9 @@
+package umc.product.domain.study.service.member;
+
+import umc.product.domain.study.entity.StudyAttendance;
+import umc.product.domain.study.entity.StudyMember;
+
+public interface StudyAttendanceQueryService {
+
+    StudyAttendance getStudyAttendance(StudyMember studyMember, int week);
+}
